src/backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
Flask-CORS==4.0.1
packaging==24.1
cryptography==43.0.1
celery==5.4.0
redis==5.1.1
requests==2.31.0
pydub==0.25.1
ffmpeg-python==0.2.0
python-dotenv==1.0.1